Added source files
Project.ipynb cleans the data set. The data_set that was cleaned so far drops all rows that involves with ios and android because it is all 0s, and drops rows with google license. Data description is Anusha's code that does basics stats and separates the data set into years, I just ran the cleaned dataset through it.